feat(website): add theme UI for landing page
