+++ packages/website/package.json
@@ -11,11 +11,17 @@
     "clean": "gatsby clean"
   },
   "dependencies": {
+    "@theme-ui/presets": "^0.3.0",
     "docz": "^2.3.2-alpha.0",
     "gatsby": "^2.24.52",
+    "gatsby-plugin-theme-ui": "^0.3.0",
     "gatsby-theme-docz": "^2.3.1-alpha.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-intime": "^0.1.0"
+    "react-intime": "^0.1.0",
+    "theme-ui": "^0.3.1"
+  },
+  "devDependencies": {
+    "@types/theme-ui": "^0.3.6"
   }
 }
