+++ packages/website/src/gatsby-plugin-theme-ui/index.tsx
@@ -1,0 +1,3 @@
+import { swiss as preset } from "@theme-ui/presets";
+
+export default preset;
