+++ packages/website/src/gatsby-plugin-theme-ui/types.d.ts
@@ -1,0 +1,8 @@
+declare module "@theme-ui/presets" {
+  import { Theme } from "theme-ui";
+  const presets: Record<
+    "base" | "deep" | "bulma" | "tailwind" | "swiss",
+    Theme
+  >;
+  export = presets;
+}
